add bsc and remove gambit
